Add new node type: GetPageNode in server and UI Add support for startNodeId in UI

+++ model/src/main/java/com/github/lzenczuk/ps/engine/Scenario.java
@@ -17,7 +17,7 @@ public class Scenario {
 
     private long offsetX = 0;
     private long offsetY = 0;
-    private long scale = 1;
+    private double scale = 1;
 
     private String executorName;
